+++ tests/conftest.py
@@ -9,14 +9,6 @@ class PiwikSettings(Testing):
     PIWIK_ID_API = 2
     PIWIK_URL = os.environ.get('PIWIK_URL', 'localhost:8080')
     PIWIK_AUTH = 'eeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee'
-    PIWIK_GOALS = {
-        'NEW_DATASET': 1,
-        'NEW_REUSE': 2,
-        'NEW_FOLLOW': 3,
-        'SHARE': 4,
-        'RESOURCE_DOWNLOAD': 5,
-        'RESOURCE_REDIRECT': 6,
-    }
     METRICS_DSN = {
         'host': 'localhost',
         'port': '8086',