+++ udata_piwik/settings.py
@@ -16,17 +16,6 @@ PIWIK_ID_API = 1
 # Authentication token from Piwik/Matomo
 PIWIK_AUTH = '<32-chars-auth-token-from-piwik>'
 
-# Piwik/Matomo Goals mapping
-# All keys are required and should exists on instance
-PIWIK_GOALS = {
-    # 'NEW_DATASET': 1,
-    # 'NEW_REUSE': 2,
-    # 'NEW_FOLLOW': 3,
-    # 'SHARE': 4,
-    # 'RESOURCE_DOWNLOAD': 5,
-    # 'RESOURCE_REDIRECT': 6,
-}
-
 # Piwik/Matomo reporting analyse timeout in seconds
 PIWIK_ANALYZE_TIMEOUT = 60 * 5